Define data_block_t as uint64_t.
Elsewhere in the codebase it is treated as a uint64_t. It is defined
here as an unsigned long long, however, which may not be the same type
as uint64_t on 64-bit platforms.

printf formatting strings need to be updated from "lld" => PRId64.

Bug: 110161494

+++ block_device.h
@@ -19,7 +19,7 @@
 #include <lk/list.h>
 #include <stdint.h>
 
-typedef unsigned long long data_block_t;
+typedef uint64_t data_block_t;
 
 /**
  * struct block_device - Block device functions and state
